feat: use swr to fetch dev icons
